requirementsv1pyt.txt
torch>=1.7.0
numpy>=1.18.0
matplotlib>=3.1.0
Pillow>=8.0.0
